Add images to README. Remove config.ini. Add code to build config.ini from scratch. Change code to have default entity. Add coldstart code to init.php.

+++ models/entity.mdl.php
@@ -11,29 +11,24 @@ class entity
 	{
 		global $cfg;
 
-		if (!file_exists('coldstart.sql')) {
-			emsg('F', "Selected entity needs the file 'coldstart.sql' to start, and it's missing.");
-			redirect('index.php');
+		$cfg['dbdata'] = $cfg['datadir'] . $cfg['app_nick'] . $post['number'] . '.sq3';
+		// the entadd.php script should already prevent this, but just in
+		// case...
+		if (file_exists($cfg['dbdata'])) {
+			emsg('F', 'A database file already exists for that entity. Aborting.');
+			return;
 		}
-
-		$ncfg = [
-			'libdir' => $cfg['libdir'],
-			'dbdriv' => 'SQLite3',
-			'dbdata' => $cfg['app_nick'] . $post['number'] . '.sq3'
-		];
 		
-		$ndb = new database($ncfg);
+		$ndb = new database($cfg);
+		coldstart($ndb);
 
-		$lines = file('coldstart.sql', FILE_IGNORE_NEW_LINES);
-		foreach ($lines as $line) {
-			$ndb->query($line);
-		}
-
+		// update config file
 		$config = file_get_contents('config/config.ini');
-		$config .= "\nentity[{$post['number']}] = {$post['name']}";
+		$config .= "\nentity[{$post['number']}] = \"{$post['name']}\"";
 		file_put_contents('config/config.ini', $config);
 
 		emsg('S', 'Database created for new entity.');
+		return;
 	}
 }
